+++ packages/react-pdf-toolkit/vite.config.ts
@@ -1,7 +1,7 @@
-import dts from 'vite-plugin-dts';
+import react from '@vitejs/plugin-react';
 import { resolve } from 'path';
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
+import dts from 'vite-plugin-dts';
 
 // https://vitejs.dev/config/
 export default defineConfig({