+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,58 @@
 # kyb-app
 
+## 0.3.2
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.3
+
+## 0.3.1
+
+### Patch Changes
+
+- bumo
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.2
+  - @ballerine/blocks@0.2.2
+  - @ballerine/common@0.9.1
+  - @ballerine/ui@0.5.1
+
+## 0.3.0
+
+### Minor Changes
+
+- 9fe7a5c10: Bump
+
+### Patch Changes
+
+- ver bump
+- 9fe7a5c10: bump
+- 9fe7a5c10: Version bump
+- Updated dependencies
+- Updated dependencies [9fe7a5c10]
+- Updated dependencies [9fe7a5c10]
+- Updated dependencies [9fe7a5c10]
+- Updated dependencies [9fe7a5c10]
+- Updated dependencies [9fe7a5c10]
+  - @ballerine/blocks@0.2.1
+  - @ballerine/common@0.9.0
+  - @ballerine/ui@0.5.0
+  - @ballerine/workflow-browser-sdk@0.6.1
+
+## 0.3.0-11210907c.0
+
+### Minor Changes
+
+- Bump
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.0-11210907c.0
+  - @ballerine/ui@0.5.0-11210907c.0
+  - @ballerine/blocks@0.2.1-11210907c.0
+  - @ballerine/workflow-browser-sdk@0.6.1-11210907c.0
+
 ## 0.2.0
 
 ### Minor Changes
@@ -102,6 +155,10 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @ballerine/common@0.7.42-11210907c.0
+  - @ballerine/blocks@0.1.27-11210907c.0
+  - @ballerine/ui@0.3.29-11210907c.0
+  - @ballerine/workflow-browser-sdk@0.5.42-11210907c.0
   - @ballerine/common@0.7.42
   - @ballerine/workflow-browser-sdk@0.5.42